fix upgrade guide snippets
