+++ Snippets/Gateway/Gateway_3/UpgradeGuide/2to3/UpgradeGuideFeatureUsage.cs
@@ -1,0 +1,16 @@
+﻿namespace Gateway_3.UpgradeGuide._2to3
+{
+    using NServiceBus;
+
+    class UpgradeGuideFeatureUsage
+    {
+        void EnableGatewayAfter(EndpointConfiguration endpointConfiguration)
+        {
+            #region 2to3EnableGatewayAfter
+
+            endpointConfiguration.Gateway();
+
+            #endregion
+        }
+    }
+}
